Don't assume a fixed header length on single-nalu
Previously we computed the size of the IDR header only once and
assumed that:

1. Every IDR frame would have the same header size
2. Only IDR frames would contain SPS and PPS

While this is very efficient, it may very easily be
incorrect. Instead, now we only identify IDR frames by the appropriate
NAL unit type, and resize the buffer to discard everything but the
actual data. This was done in a generic way, such that even non-IDR
slices can also be resized, if they contain SPS and PPS.